+++ src/pages/chat/components/sidebar/setting-tap/channel-property-content.tsx
@@ -1,0 +1,5 @@
+const ChannelPropertyContent = (props: { label: string }) => {
+	return <div className='text-sm font-normal gray-900'>{props.label}</div>;
+};
+
+export default ChannelPropertyContent;
