+++ src/pages/chat/components/sidebar/setting-tap/channel-property-title.tsx
@@ -1,0 +1,5 @@
+const ChannelPropertyTitle = (props: { label: string }) => {
+	return <div className='text-sm font-bold'>{props.label}</div>;
+};
+
+export default ChannelPropertyTitle;
